requirements_test.txt
# App requirements from setup.py
LatLon23>=1,<2
Pillow>=3.4.2,<3.5
django-auth-ldap>=1.2.8,<2
django-constance[database]>=1.3,<2
django-filter<1
django-geoposition>=0.3
django-haystack>=2.5.1,<2.7
django-import-export>=0.5.1,<1
django-modeltranslation>=0.12,<0.13
django-rest-auth>=0.8.1,<1
django-rest-swagger>=2.1.0,<2.2
django-rq>=0.9.4,<1
django-taggit-serializer>=0.1.5,<0.2
django-taggit>=0.21.3,<0.23
django-thumbor>=0.5.6,<0.6
django-yubin>=0.3.1,<1
django_categories>=1.4.3,<1.5
djangorestframework-recursive>=0.1.1,<0.2
djangorestframework>=3.5.2,<3.7
drf-chunked-upload==0.2.2
drf-haystack>=1.6.0,<1.7
dry-rest-permissions>=0.1.8,<0.2
exifread>=2.1.2,<3
flickrapi>=2.1.2,<3
hashfs>=0.7.0,<0.8
python-dateutil>=2.5.3,<3
rest-framework-generic-relations>=1.1.0,<1.2

# Additional requirements for development and testing
Django>=1.10,<1.11
faker==0.7.3
flake8==3.3.0
ipdb==0.10.2
model_mommy==1.3.0
pytest==3.0.7
pytest-django==3.1.2
pytest-sugar==0.8.0
tox==2.7.0

# local bima_core
-e .
